feat(question_type): defining schemas for question type
