+++ src/question_type/repository.py
@@ -1,0 +1,2 @@
+from src.database.repository import Repository
+
